Rename EmailAutofill class
